+++ examples/importsculib.py
@@ -1,3 +1,0 @@
-from disq.sculib import SCU
-
-print(SCU)